minor fix for database schema conflict.

+++ src/TicketPlatform.Core/Common/BaseEntity.cs
@@ -6,5 +6,5 @@ public abstract class BaseEntity
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     public DateTime? UpdatedAt { get; set; }
 
-    public uint Version { get; set; }
+    public long  Version { get; set; }
 }
